# EICR_App/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CertMate (EICR-oMatic 3000) API
  version: 1.0.0
  description: |
    Backend API for CertMate — automated EICR/EIC electrical certificate creation
    for electrical inspectors. Powered by Express/Node.js with ES modules.

    ## Authentication
    Most endpoints require a Bearer JWT token obtained via `POST /api/auth/login`.
    Pass the token in the `Authorization` header: `Bearer <token>`.

    ## Certificate Types
    - **EICR** — Electrical Installation Condition Report (periodic inspection)
    - **EIC** — Electrical Installation Certificate (new installations)
  contact:
    name: CertMate Support
  license:
    name: Proprietary
servers:
  - url: https://certomatic3000.co.uk
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
...
security:
  - BearerAuth: []
paths:
...
  /api/settings/{userId}/defaults:
    get:
      tags: [Settings]
      summary: Get user circuit field defaults
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User defaults object
          content:
            application/json:
              schema:
                type: object
    put:
      tags: [Settings]
      summary: Update user circuit field defaults
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Defaults updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
  /api/settings/{userId}/company:
    get:
      tags: [Settings]
      summary: Get company settings
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  company_name:
                    type: string
                  company_address:
                    type: string
                  company_phone:
                    type: string
                  company_email:
                    type: string
                  company_website:
                    type: string
                  company_registration:
                    type: string
                  logo_file:
                    type: string
                    nullable: true
    put:
      tags: [Settings]
      summary: Update company settings
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
  /api/inspector-profiles/{userId}:
    get:
      tags: [Settings]
      summary: Get inspector profiles for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of inspector profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    put:
      tags: [Settings]
      summary: Update inspector profiles
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
      responses:
        "200":
          description: Profiles updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
  /api/inspector-profiles/{userId}/upload-signature:
    post:
      tags: [Settings]
      summary: Upload inspector signature image
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [signature]
              properties:
                signature:
                  type: string
                  format: binary
      responses:
        "200":
          description: Signature uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  signature_file:
                    type: string
...
  /api/regulations:
    get:
      tags: [Settings]
      summary: Search BS 7671 regulations
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search term
      responses:
        "200":
          description: Matching regulations (max 20)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ref:
                      type: string
                    section:
                      type: string
                    title:
                      type: string
                    description:
                      type: string

  # ============= Clients =============
...
  /api/push/vapid-key:
    get:
      tags: [Push]
      summary: Get VAPID public key for push subscriptions
      security: []
      responses:
        "200":
          description: VAPID key
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
        "503":
          description: Push not configured
  /api/push/subscribe:
    post:
      tags: [Push]
      summary: Subscribe to push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscription]
              properties:
                subscription:
                  type: object
                  required: [endpoint, keys]
                  properties:
                    endpoint:
                      type: string
                      format: uri
                    keys:
                      type: object
                      required: [p256dh, auth]
                      properties:
                        p256dh:
                          type: string
                        auth:
                          type: string
      responses:
        "200":
          description: Subscription saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Invalid subscription
  /api/push/unsubscribe:
    post:
      tags: [Push]
      summary: Unsubscribe from push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint:
                  type: string
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Endpoint required

  # ============= Keys & Proxy =============
...